fix:린트정렬 수정, 추천 명소 폴더 홈폴더 하위로 재배치
