+++ lib/views/home/recommended_place/recommended_place_detail_page.dart
@@ -1,25 +1,30 @@
 import 'package:flutter/material.dart';
-import 'package:travel_muse_app/views/recommended_place/widgets/action_button_row.dart';
-import 'package:travel_muse_app/views/recommended_place/widgets/image_slider.dart';
-import 'package:travel_muse_app/views/recommended_place/widgets/location_row.dart';
-import 'package:travel_muse_app/views/recommended_place/widgets/place_direction_info.dart';
-import 'package:travel_muse_app/views/recommended_place/widgets/place_map_view.dart';
-import 'package:travel_muse_app/views/recommended_place/widgets/place_stats_row.dart';
+import 'package:travel_muse_app/views/home/recommended_place/widgets/action_button_row.dart';
+import 'package:travel_muse_app/views/home/recommended_place/widgets/image_slider.dart';
+import 'package:travel_muse_app/views/home/recommended_place/widgets/location_row.dart';
+import 'package:travel_muse_app/views/home/recommended_place/widgets/place_description.dart';
+import 'package:travel_muse_app/views/home/recommended_place/widgets/place_direction_info.dart';
+import 'package:travel_muse_app/views/home/recommended_place/widgets/place_info_section.dart';
+import 'package:travel_muse_app/views/home/recommended_place/widgets/place_map_view.dart';
+import 'package:travel_muse_app/views/home/recommended_place/widgets/place_stats_row.dart';
 
-class RecommendedRestaurantDetailPage extends StatefulWidget {
-  const RecommendedRestaurantDetailPage({super.key});
+class RecommendedPlaceDetailPage extends StatefulWidget {
+  const RecommendedPlaceDetailPage({super.key});
 
   @override
-  State<RecommendedRestaurantDetailPage> createState() =>
-      _RecommendedRestaurantDetailPageState();
+  State<RecommendedPlaceDetailPage> createState() =>
+      _RecommendedPlaceDetailPageState();
 }
 
-class _RecommendedRestaurantDetailPageState
-    extends State<RecommendedRestaurantDetailPage> {
+class _RecommendedPlaceDetailPageState
+    extends State<RecommendedPlaceDetailPage> {
   final Color primaryColor = const Color(0xFF03A9F4);
   final Color secondaryColor = const Color(0xFF1E88E5);
 
-  final List<String> imageUrls = ['assets/images/image3.png'];
+  final List<String> imageUrls = [
+    'assets/images/image1.png',
+    'assets/images/image2.png',
+  ];
 
   int _currentPage = 0;
   final PageController _pageController = PageController();
@@ -36,11 +41,11 @@ class _RecommendedRestaurantDetailPageState
           icon: const Icon(Icons.arrow_back),
           onPressed: () => Navigator.pop(context),
         ),
-        title: const Text('핫도그 두물머리점'),
+        title: const Text('두물머리'),
       ),
       floatingActionButton: FloatingActionButton.extended(
         backgroundColor: secondaryColor,
-        icon: const Icon(Icons.restaurant),
+        icon: const Icon(Icons.navigation),
         label: const Text('길찾기'),
         onPressed: () {
           // 위치 서비스 연동 예정
@@ -51,6 +56,7 @@ class _RecommendedRestaurantDetailPageState
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
+            // 상단 이미지 슬라이더
             ImageSlider(
               imageUrls: imageUrls,
               currentPage: _currentPage,
@@ -58,26 +64,23 @@ class _RecommendedRestaurantDetailPageState
               onPageChanged: (index) => setState(() => _currentPage = index),
             ),
             const SizedBox(height: 16),
+
             const PlaceStatsRow(),
             const SizedBox(height: 12),
+
             const LocationRow(),
             const SizedBox(height: 24),
+
             const ActionButtonRow(),
             const SizedBox(height: 24),
 
-            // 다른 설명만 바뀜
-            const Text(
-              '두물머리 인증샷 명소 바로 앞! 바삭하고 치즈가득한 수제 핫도그로 유명한 곳.',
-              style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-            ),
+            const PlaceDescription(),
             const SizedBox(height: 24),
 
             const PlaceMapView(),
             const SizedBox(height: 16),
 
-            const Text('주소: 경기 양평군 양서면 두물머리길'),
-            const Text('전화: 031-000-0000'),
-            const Text('홈페이지: 없음', style: TextStyle(color: Colors.grey)),
+            const PlaceInfoSection(),
 
             const PlaceDirectionInfo(),
