+++ lib/views/home/recommended_place/widgets/image_slider.dart
