+++ lib/views/home/recommended_place/widgets/location_row.dart
