+++ lib/views/home/recommended_place/widgets/place_info_section.dart
